requirements.txt
azure-identity==1.7.1
azure-mgmt-containerinstance==8.0.0